refactor: `types`の格納場所を共通化
