+++ app/src/utils/gol-patterns/clock-02.ts
@@ -1,5 +1,5 @@
 import { ALIVE as A, DEAD as D } from "../types/cell-state";
-import type { PatternConfigs } from "./types";
+import type { PatternConfigs } from "../types/pattern-configs";
 
 // 時計II
 export const clock02: PatternConfigs = {
