+++ app/src/utils/gol-patterns/glider-gun.ts
@@ -1,5 +1,5 @@
 import { ALIVE as A, DEAD as D } from "../types/cell-state";
-import type { PatternConfigs } from "./types";
+import type { PatternConfigs } from "../types/pattern-configs";
 
 // グライダー銃
 export const gliderGun: PatternConfigs = {
