+++ app/src/utils/gol-patterns/middleweight-spaceship.ts
@@ -1,5 +1,5 @@
 import { ALIVE as A, DEAD as D } from "../types/cell-state";
-import type { PatternConfigs } from "./types";
+import type { PatternConfigs } from "../types/pattern-configs";
 
 // 中量級宇宙船
 export const middleweightSpaceship: PatternConfigs = {
